+++ dandelion/util.py
@@ -3,6 +3,7 @@ import numpy as np
 import theano
 import theano.tensor as tensor
 import gzip, pickle
+import socket, time
 
 def floatX(arr):
     """Converts data to a numpy array of dtype ``theano.config.floatX``.
@@ -493,8 +494,8 @@ class gpickle(object):
     A pickle class with gzip enabled
     """
     @staticmethod
-    def dump(data, filename):
-        with gzip.open(filename, mode='wb') as f:
+    def dump(data, filename, compresslevel=9):
+        with gzip.open(filename, mode='wb', compresslevel=compresslevel) as f:
             pickle.dump(data, chunked_byte_writer(f))
             f.close()
     @staticmethod
@@ -568,3 +569,109 @@ def theano_safe_run(fn, input_list):
             return status, None
         else:
             raise e
+
+class Finite_Memory_Array(object):
+
+    def __init__(self, array=None, shape=None, dtype=np.double):
+        if array is not None:
+            self.array = array
+        elif shape is not None:
+            self.array = np.zeros(shape=shape, dtype=dtype)
+        else:
+            raise ValueError('bad initialization para for Finite_Memory_Array')
+        self.shape = self.array.shape
+        self.curpos = 0
+        self.first_round = True
+
+    def update(self, value):
+        self.array[:, self.curpos] = value
+        self.curpos += 1
+        if self.curpos >= self.shape[1]:
+            self.curpos = 0
+            self.first_round = False
+
+    def clear(self):
+        self.array[:] = 0
+        self.curpos = 0
+        self.first_round = True
+
+    def get_current_position(self):
+        return self.curpos
+
+    @property
+    def content(self):
+        if self.first_round is True:
+            return self.array[:, :self.curpos]
+        else:
+            return self.array
+
+def pad_sequence_into_array(Xs, maxlen=None, truncating='post', padding='post', value=0.):
+    """
+    Padding sequence (list of numpy arrays) into an numpy array
+    :param Xs: list of numpy arrays. The arrays must have the same shape except the first dimension.
+    :param maxlen: the allowed maximum of the first dimension of Xs's arrays. Any array longer than maxlen is truncated to maxlen
+    :param truncating: = 'pre'/'post', indicating whether the truncation happens at either the beginning or the end of the array (default)
+    :param padding: = 'pre'/'post',indicating whether the padding happens at either the beginning or the end of the array (default)
+    :param value: scalar, the padding value, default = 0.0
+    :return: Xout, the padded sequence (now an augmented array with shape (Narrays, N1stdim, N2nddim, ...)
+    :return: mask, the corresponding mask, binary array, with shape (Narray, N1stdim)
+    """
+    Nsamples = len(Xs)
+    if maxlen is None:
+        lengths = [s.shape[0] for s in Xs]    # 'sequences' must be list, 's' must be numpy array, len(s) return the first dimension of s
+        maxlen = np.max(lengths)
+
+    Xout = np.ones(shape = [Nsamples, maxlen] + list(Xs[0].shape[1:]), dtype=Xs[0].dtype) * np.asarray(value, dtype=Xs[0].dtype)
+    Mask = np.zeros(shape = [Nsamples, maxlen], dtype=Xout.dtype)
+    for i in range(Nsamples):
+        x = Xs[i]
+        if truncating == 'pre':
+            trunc = x[-maxlen:]
+        elif truncating == 'post':
+            trunc = x[:maxlen]
+        else:
+            raise ValueError("Truncating type '%s' not understood" % truncating)
+        if padding == 'post':
+            Xout[i, :len(trunc)] = trunc
+            Mask[i, :len(trunc)] = 1
+        elif padding == 'pre':
+            Xout[i, -len(trunc):] = trunc
+            Mask[i, -len(trunc):] = 1
+        else:
+            raise ValueError("Padding type '%s' not understood" % padding)
+    return Xout, Mask
+
+def get_local_ip():
+    """
+    Get local host IP address
+    :return:
+    """
+    s = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
+    s.connect(('8.8.8.8', 0))                                # connecting to a UDP address doesn't send packets
+    local_ip_address = s.getsockname()[0]
+    return local_ip_address
+
+def get_time_stamp():
+    """
+    Create a formatted string time stamp
+    :return:
+    """
+    return time.strftime("%Y-%m-%d_%H-%M-%S", time.localtime())
+
+
+if __name__ == '__main__':
+    INFO = ['This is a collection of auxiliary functions for DL.\n',
+            'Author: David Leon\n',
+            'All rights reserved\n']
+    print(*INFO)
+
+    # x1 = np.array([1,2,3])
+    # x2 = np.array([1,2,3,4])
+    # x3 = np.array([5,4,3,2,1])
+    x1, x2, x3 = np.random.rand(2, 4), np.random.rand(3, 4), np.random.rand(5,4)
+
+    Xout, mask = pad_sequence_into_array([x1,x2,x3])
+    print(Xout)
+    print(mask)
+    print(Xout.shape)
+    print(mask.shape)